exp: Rename jaeger-v2 binary to just jaeger
