--- cmd/jaeger-v2/.gitignore
@@ -1,2 +1,0 @@
-jaeger-v2-*-*
-jaeger-v2
