+++ cmd/jaeger/.gitignore
@@ -1,0 +1,2 @@
+jaeger-*-*
+jaeger
