+++ cmd/jaeger/README.md
@@ -1,0 +1,4 @@
+# jaeger
+
+This is experimental Jaeger V2 based on OpenTelemetry collector.
+See https://github.com/jaegertracing/jaeger/issues/4843.
