+++ cmd/jaeger/config-ui.json
