+++ cmd/jaeger/internal/extension/jaegerquery/README.md
