+++ cmd/jaeger/internal/extension/jaegerquery/factory.go
