+++ cmd/jaeger/internal/extension/jaegerstorage/factory.go
